Rebase branch

+++ packages/flutter_tools/test/src/common.dart
@@ -4,23 +4,28 @@
 
 import 'dart:async';
 
+import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
+import 'package:flutter_tools/src/base/logger.dart';
+import 'package:flutter_tools/src/base/platform.dart';
 import 'package:flutter_tools/src/convert.dart';
 import 'package:vm_service/vm_service.dart' as vm_service;
+import 'package:path/path.dart' as path; // ignore: package_path_import
 
 import 'package:flutter_tools/src/base/common.dart';
 import 'package:flutter_tools/src/base/context.dart';
 import 'package:flutter_tools/src/base/file_system.dart';
-import 'package:flutter_tools/src/base/process.dart';
+import 'package:flutter_tools/src/base/io.dart';
 import 'package:flutter_tools/src/commands/create.dart';
 import 'package:flutter_tools/src/runner/flutter_command.dart';
 import 'package:flutter_tools/src/runner/flutter_command_runner.dart';
 import 'package:flutter_tools/src/globals.dart' as globals;
 import 'package:meta/meta.dart';
+import 'package:fake_async/fake_async.dart';
 import 'package:test_api/test_api.dart' as test_package show TypeMatcher, test; // ignore: deprecated_member_use
 import 'package:test_api/test_api.dart' hide TypeMatcher, isInstanceOf; // ignore: deprecated_member_use
 // ignore: deprecated_member_use
-export 'package:test_core/test_core.dart' hide TypeMatcher, isInstanceOf; // Defines a 'package:test' shim.
+export 'package:test_core/test_core.dart' hide TypeMatcher, isInstanceOf, test; // Defines a 'package:test' shim.
 
 /// A matcher that compares the type of the actual value to the type argument T.
 // TODO(ianh): Remove this once https://github.com/dart-lang/matcher/issues/98 is fixed
@@ -31,7 +36,9 @@ void tryToDelete(Directory directory) {
   // on Windows it's common for deletions to fail due to
   // bogus (we think) "access denied" errors.
   try {
-    directory.deleteSync(recursive: true);
+    if (directory.existsSync()) {
+      directory.deleteSync(recursive: true);
+    }
   } on FileSystemException catch (error) {
     print('Failed to delete ${directory.path}: $error');
   }
@@ -43,8 +50,9 @@ void tryToDelete(Directory directory) {
 /// environment variable is set, it will be returned. Otherwise, this will
 /// deduce the path from `platform.script`.
 String getFlutterRoot() {
-  if (globals.platform.environment.containsKey('FLUTTER_ROOT')) {
-    return globals.platform.environment['FLUTTER_ROOT'];
+  const Platform platform = LocalPlatform();
+  if (platform.environment.containsKey('FLUTTER_ROOT')) {
+    return platform.environment['FLUTTER_ROOT'];
   }
 
   Error invalidScript() => StateError('Could not determine flutter_tools/ path from script URL (${globals.platform.script}); consider setting FLUTTER_ROOT explicitly.');
@@ -66,31 +74,21 @@ String getFlutterRoot() {
       throw invalidScript();
   }
 
-  final List<String> parts = globals.fs.path.split(globals.fs.path.fromUri(scriptUri));
+  final List<String> parts = path.split(globals.fs.path.fromUri(scriptUri));
   final int toolsIndex = parts.indexOf('flutter_tools');
   if (toolsIndex == -1) {
     throw invalidScript();
   }
-  final String toolsPath = globals.fs.path.joinAll(parts.sublist(0, toolsIndex + 1));
-  return globals.fs.path.normalize(globals.fs.path.join(toolsPath, '..', '..'));
+  final String toolsPath = path.joinAll(parts.sublist(0, toolsIndex + 1));
+  return path.normalize(path.join(toolsPath, '..', '..'));
 }
 
 CommandRunner<void> createTestCommandRunner([ FlutterCommand command ]) {
-  final FlutterCommandRunner runner = FlutterCommandRunner();
+  final FlutterCommandRunner runner = TestFlutterCommandRunner();
   if (command != null) {
     runner.addCommand(command);
   }
   return runner;
-}
-
-/// Updates [path] to have a modification time [seconds] from now.
-void updateFileModificationTime(
-  String path,
-  DateTime baseTime,
-  int seconds,
-) {
-  final DateTime modificationTime = baseTime.add(Duration(seconds: seconds));
-  globals.fs.file(path).setLastModifiedSync(modificationTime);
 }
 
 /// Matcher for functions that throw [AssertionError].
@@ -111,15 +109,17 @@ Matcher throwsToolExit({ int exitCode, Pattern message }) {
 /// Matcher for [ToolExit]s.
 final test_package.TypeMatcher<ToolExit> isToolExit = isA<ToolExit>();
 
-/// Matcher for functions that throw [ProcessExit].
-Matcher throwsProcessExit([ dynamic exitCode ]) {
-  return exitCode == null
-      ? throwsA(isProcessExit)
-      : throwsA(allOf(isProcessExit, (ProcessExit e) => e.exitCode == exitCode));
+/// Matcher for functions that throw [ProcessException].
+Matcher throwsProcessException({ Pattern message }) {
+  Matcher matcher = isProcessException;
+  if (message != null) {
+    matcher = allOf(matcher, (ProcessException e) => e.message?.contains(message));
+  }
+  return throwsA(matcher);
 }
 
-/// Matcher for [ProcessExit]s.
-final test_package.TypeMatcher<ProcessExit> isProcessExit = isA<ProcessExit>();
+/// Matcher for [ProcessException]s.
+final test_package.TypeMatcher<ProcessException> isProcessException = isA<ProcessException>();
 
 /// Creates a flutter project in the [temp] directory using the
 /// [arguments] list if specified, or `--no-pub` if not.
@@ -156,6 +156,35 @@ Matcher containsIgnoringWhitespace(String toSearch) {
   );
 }
 
+/// The tool overrides `test` to ensure that files created under the
+/// system temporary directory are deleted after each test by calling
+/// `LocalFileSystem.dispose()`.
+@isTest
+void test(String description, FutureOr<void> body(), {
+  String testOn,
+  Timeout timeout,
+  dynamic skip,
+  List<String> tags,
+  Map<String, dynamic> onPlatform,
+  int retry,
+}) {
+  test_package.test(
+    description,
+    () async {
+      addTearDown(() async {
+        await LocalFileSystem.dispose();
+      });
+      return body();
+    },
+    timeout: timeout,
+    skip: skip,
+    tags: tags,
+    onPlatform: onPlatform,
+    retry: retry,
+    testOn: testOn,
+  );
+}
+
 /// Executes a test body in zone that does not allow context-based injection.
 ///
 /// For classes which have been refactored to excluded context-based injection
@@ -168,12 +197,12 @@ Matcher containsIgnoringWhitespace(String toSearch) {
 void testWithoutContext(String description, FutureOr<void> body(), {
   String testOn,
   Timeout timeout,
-  bool skip,
+  dynamic skip,
   List<String> tags,
   Map<String, dynamic> onPlatform,
   int retry,
   }) {
-  return test_package.test(
+  return test(
     description, () async {
       return runZoned(body, zoneValues: <Object, Object>{
         contextKey: const NoContext(),
@@ -186,6 +215,21 @@ void testWithoutContext(String description, FutureOr<void> body(), {
     retry: retry,
     testOn: testOn,
   );
+}
+
+/// Runs a callback using FakeAsync.run while continually pumping the
+/// microtask queue. This avoids a deadlock when tests `await` a Future
+/// which queues a microtask that will not be processed unless the queue
+/// is flushed.
+Future<T> runFakeAsync<T>(Future<T> Function(FakeAsync time) f) async {
+  return FakeAsync().run((FakeAsync time) async {
+    bool pump = true;
+    final Future<T> future = f(time).whenComplete(() => pump = false);
+    while (pump) {
+      time.flushMicrotasks();
+    }
+    return future;
+  });
 }
 
 /// An implementation of [AppContext] that throws if context.get is called in the test.
@@ -239,7 +283,6 @@ class FakeVmServiceHost {
       final FakeVmServiceRequest fakeRequest = _requests.removeAt(0) as FakeVmServiceRequest;
       expect(request, isA<Map<String, Object>>()
         .having((Map<String, Object> request) => request['method'], 'method', fakeRequest.method)
-        .having((Map<String, Object> request) => request['id'], 'id', fakeRequest.id)
         .having((Map<String, Object> request) => request['params'], 'args', fakeRequest.args)
       );
       if (fakeRequest.close) {
@@ -250,13 +293,13 @@ class FakeVmServiceHost {
       if (fakeRequest.errorCode == null) {
         _input.add(json.encode(<String, Object>{
           'jsonrpc': '2.0',
-          'id': fakeRequest.id,
+          'id': request['id'],
           'result': fakeRequest.jsonResponse ?? <String, Object>{'type': 'Success'},
         }));
       } else {
         _input.add(json.encode(<String, Object>{
           'jsonrpc': '2.0',
-          'id': fakeRequest.id,
+          'id': request['id'],
           'error': <String, Object>{
             'code': fakeRequest.errorCode,
           }
@@ -299,15 +342,13 @@ abstract class VmServiceExpectation {
 class FakeVmServiceRequest implements VmServiceExpectation {
   const FakeVmServiceRequest({
     @required this.method,
-    @required this.id,
-    @required this.args,
+    this.args = const <String, Object>{},
     this.jsonResponse,
     this.errorCode,
     this.close = false,
   });
 
   final String method;
-  final String id;
 
   /// When true, the vm service is automatically closed.
   final bool close;
@@ -333,4 +374,42 @@ class FakeVmServiceStreamResponse implements VmServiceExpectation {
 
   @override
   bool get isRequest => false;
+}
+
+class TestFlutterCommandRunner extends FlutterCommandRunner {
+  @override
+  Future<void> runCommand(ArgResults topLevelResults) async {
+    final Logger topLevelLogger = globals.logger;
+    final Map<Type, dynamic> contextOverrides = <Type, dynamic>{
+      if (topLevelResults['verbose'] as bool)
+        Logger: VerboseLogger(topLevelLogger),
+    };
+    return context.run<void>(
+      overrides: contextOverrides.map<Type, Generator>((Type type, dynamic value) {
+        return MapEntry<Type, Generator>(type, () => value);
+      }),
+      body: () => super.runCommand(topLevelResults),
+    );
+  }
+}
+
+/// A file system that allows preconfiguring certain entities.
+///
+/// This is useful for inserting mocks/entities which throw errors or
+/// have other behavior that is not easily configured through the
+/// filesystem interface.
+class ConfiguredFileSystem extends ForwardingFileSystem {
+  ConfiguredFileSystem(FileSystem delegate, {@required this.entities}) : super(delegate);
+
+  final Map<String, FileSystemEntity> entities;
+
+  @override
+  File file(dynamic path) {
+    return (entities[path] as File) ?? super.file(path);
+  }
+
+  @override
+  Directory directory(dynamic path) {
+    return (entities[path] as Directory) ?? super.directory(path);
+  }
 }
